+++ pyxel/models/charge_generation/__init__.py
@@ -6,3 +6,10 @@
 #  the terms contained in the file ‘LICENCE.txt’.
 
 """TBW."""
+
+# flake8: noqa
+from .charge_injection import charge_blocks
+from .dark_current import dark_current_rule07
+from .load_profile import charge_profile
+from .photoelectrons import monte_carlo_conversion, simple_conversion, random_pos
+from .tars.tars import run_tars